+++ src/models/src/NIRadioGroup.m
@@ -77,7 +77,7 @@ static const NSInteger kInvalidSelection = NSIntegerMin;
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (id)keyForIdentifier:(NSInteger)identifier {
-  return [NSNumber numberWithInt:identifier];
+  return [NSNumber numberWithInt:((int)identifier)];
 }
